Created AIData class; AI now derives from Controller; Updated classes to use doubles instead of floats; Added forward declaration header ai.h;

+++ src/game/ai/ailogicblock.h
@@ -11,7 +11,7 @@ class AILogicBlock {
 	~AILogicBlock() {}
 
 	virtual void Start() {}
-	virtual AIModule::Status Update(float dt) { return AIModule::DONE; }
+	virtual AIModule::Status Update(double dt) { return AIModule::DONE; }
 	virtual void Finish() {}
 
 	LogicModule* get_parent() { return parent_; }
